+++ src/Kinepolis/Parser/PriceParser.php
@@ -2,7 +2,9 @@
 
 declare(strict_types=1);
 
-namespace CultuurNet\UDB3\Kinepolis;
+namespace CultuurNet\UDB3\Kinepolis\Parser;
+
+use CultuurNet\UDB3\Kinepolis\ParsedPriceForATheater;
 
 interface PriceParser
 {
